440: Split candidates controller into separate files for each use case
